changed the type notation for the argument

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -8,7 +8,7 @@ from typing import List
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: int, max_delay: float = 10) -> List[float]:
+async def wait_n(n: int, max_delay: int = 10) -> List[float]:
     """
     Args:
         n (int): num of time to spawn wait_random.
@@ -17,5 +17,6 @@ async def wait_n(n: int, max_delay: float = 10) -> List[float]:
     Returns:
         list[float]: List of delays in ascending order
     """
+
     values = [asyncio.create_task(wait_random(max_delay)) for _ in range(n)]
     return [await task for task in asyncio.as_completed(values)]
